Base TimeEntity 생성

+++ back/src/main/java/com/sun/back/dto/diaryGroup/GetMyGroupResponse.java
@@ -8,5 +8,5 @@ public record GetMyGroupResponse(
         String title,
         MemberRole role,
         GroupType groupType,
-        Long membersCount
+        Integer membersCount
 ){}
